refactor: 프레젠테이션 계층 DTO 금액 타입 Long → BigDecimal 변경
- Request DTO 6종: @Positive → @DecimalMin("0.01") + @Digits(integer=19, fraction=2)
- Response DTO 14종: 금액 필드 Long → BigDecimal 변경

+++ payment/src/main/java/com/example/payment/presentation/dto/response/PaymentRefundItemResponse.java
@@ -2,12 +2,13 @@ package com.example.payment.presentation.dto.response;
 
 import com.example.payment.application.dto.PaymentRefundItemResult;
 import com.example.payment.domain.enumtype.PaymentRefundItemStatus;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record PaymentRefundItemResponse(
         UUID orderItemId,
         PaymentRefundItemStatus status,
-        Long refundAmount
+        BigDecimal refundAmount
 ) {
     public static PaymentRefundItemResponse from(PaymentRefundItemResult result) {
         return new PaymentRefundItemResponse(
