+++ payment/src/main/java/com/example/payment/presentation/dto/response/WithdrawResponse.java
@@ -2,17 +2,18 @@ package com.example.payment.presentation.dto.response;
 
 import com.example.payment.application.dto.WithdrawResult;
 import com.example.payment.domain.enumtype.WithdrawStatus;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 public record WithdrawResponse(
         UUID withdrawRequestId,
-        Long amount,
-        Long fee,
-        Long actualAmount,
+        BigDecimal amount,
+        BigDecimal fee,
+        BigDecimal actualAmount,
         String maskedBankAccount,
         WithdrawStatus status,
-        Long walletBalance,
+        BigDecimal walletBalance,
         LocalDateTime requestedAt,
         LocalDateTime processedAt
 ) {
